requirements.txt
# ─── Core ML ────────────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0

# ─── RLHF / TRL ─────────────────────────────────────────────────────────────
trl>=0.8.6

# ─── Gymnasium (custom text-gen env) ─────────────────────────────────────────
gymnasium>=0.29.1

# ─── MLOps ───────────────────────────────────────────────────────────────────
mlflow>=2.12.0

# ─── API ─────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ─── Evaluation ──────────────────────────────────────────────────────────────
evaluate>=0.4.1
rouge-score>=0.1.2
nltk>=3.8.1
sacrebleu>=2.4.0

# ─── Config & Utilities ──────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.1
tqdm>=4.66.2
numpy>=1.26.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.1.0
httpx>=0.27.0    # async test client for FastAPI
